+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 * Update messages written to stdout by the `delete-my-tweets` and
  `delete-my-likes` commands.
 
+* Add `--no-links` option to the `archive-timeline` command.
+
+* Add `--no-media` option to the `archive-timeline` command.
+
 * Add `--outbound-links-only` option to the `archive-timeline` command.
 
 ## v0.1.0